editiontree: use qreals for size and positioning
* Use qreals in computeDimensions() and draw()
* Use QPointF, QRectF and QFontMetricsF in these methods
* When there are new attributes (eg. Sigma), make them qreal as well

+++ lib/symbolibre/editiontree/EditionArea.hpp
@@ -74,8 +74,8 @@ public:
     }
 
     void computeDimensions(QPainter &painter,
-                           int lheight, int lascent) override;
-    void draw(int x, int y, QPainter &painter, bool cursor) override;
+                           qreal lheight, qreal lascent) override;
+    void draw(qreal x, qreal y, QPainter &painter, bool cursor) override;
 };
 
 #endif // EDITIONAREA_HPP
