+++ lib/symbolibre/editiontree/Frac.hpp
@@ -35,8 +35,8 @@ public:
 
     /* About computing dimensions */
     void computeDimensions(QPainter &painter,
-                           int lheight, int lascent) override;
-    void draw(int x, int y, QPainter &painter, bool cursor) override;
+                           qreal lheight, qreal lascent) override;
+    void draw(qreal x, qreal y, QPainter &painter, bool cursor) override;
 };
 
 #endif // FRAC_HPP
